text: Add support for quoted GNUStep extended values
GNUStep is very lax when it comes to parsing these. Closing quotes
aren't required (as pointed out by @Artoria2e5, and held up by empirical
validation), so all we need to do is strip a closing " even if we don't
detect an opening one.

This commit also fleshes out the invalid test cases for <* and reworks
some of the error-handling logic to make the errors clearer.  Type
checks have been lifted up to the head of the ext. val. parser and the
emptiness check has been tidied up a bit.

Co-authored-by: Mingye Wang <arthur200126@gmail.com>

Closes https://github.com/DHowett/go-plist/pull/51.